% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_intersection.R
\name{dst_intersection}
\alias{dst_intersection}
\title{Intersect geometries (DuckDB-backed, sf-compatible)}
\usage{
dst_intersection(
  x,
  y,
  duckdb_threads = getOption("ducksf.intersection.threads", NULL),
  prefilter_bbox = TRUE
)
}
\description{
Intersect geometries (DuckDB-backed, sf-compatible)
}
