% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_export.R
\name{ducksf_apply_keep_na}
\alias{ducksf_apply_keep_na}
\title{Apply keep_NA semantics (drop targets with all-NA on requested vars)}
\usage{
ducksf_apply_keep_na(out, output = c("sf", "tibble"), vars_used, keep_NA)
}
\description{
Apply keep_NA semantics (drop targets with all-NA on requested vars)
}
\keyword{internal}
