% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_export.R
\name{ducksf_build_interpolated_sf_view}
\alias{ducksf_build_interpolated_sf_view}
\title{Make sf-ready view (tid + geometry + interpolated cols)}
\usage{
ducksf_build_interpolated_sf_view(
  con,
  target_tbl = "target_tbl",
  ia_view = "interpolated_all",
  out_view = "interpolated_sf"
)
}
\description{
Make sf-ready view (tid + geometry + interpolated cols)
}
\keyword{internal}
