% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_relate.R
\name{ducksf_build_overlap}
\alias{ducksf_build_overlap}
\title{Build a skinny overlap relation between projected source/target views}
\usage{
ducksf_build_overlap(
  con,
  src_view = "source_proj",
  trg_view = "target_proj",
  out_view = "overlap",
  join_pred = "intersects",
  measure = c("area", "length", "count"),
  keep_geom = FALSE,
  materialize = FALSE,
  dwithin_distance = NULL,
  prefilter_bbox = TRUE,
  src_id_col = "sid",
  trg_id_col = "tid",
  geom_col_name = "geom"
)
}
\description{
Build a skinny overlap relation between projected source/target views
}
\keyword{internal}
