% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_views.R
\name{ducksf_build_proj_views}
\alias{ducksf_build_proj_views}
\title{Build normalized + (optionally) projected source/target views}
\usage{
ducksf_build_proj_views(
  con,
  sidQN,
  tidQN,
  src_lit,
  trg_lit,
  join_lit,
  same_src_join,
  same_trg_join,
  vars = character(0),
  source_tbl_raw = "source_tbl_raw",
  target_tbl_raw = "target_tbl_raw",
  src_geom_name = "geometry",
  trg_geom_name = "geometry"
)
}
\description{
Build normalized + (optionally) projected source/target views
}
\keyword{internal}
