% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ducksf_args_check.R
\name{ducksf_check_cols_present}
\alias{ducksf_check_cols_present}
\title{Assert required columns exist (simple, reusable)}
\usage{
ducksf_check_cols_present(df, cols, where = "data frame")
}
\description{
Assert required columns exist (simple, reusable)
}
\keyword{internal}
