% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_export.R
\name{ducksf_collect_output}
\alias{ducksf_collect_output}
\title{Collect result as sf or tibble, mirroring current behavior}
\usage{
ducksf_collect_output(
  con,
  output = c("sf", "tibble"),
  target_sf,
  target_crs,
  tidQN,
  nameConflict = FALSE,
  tidOrig = NULL,
  ia_view = "interpolated_all",
  isf_view = "interpolated_sf"
)
}
\description{
Collect result as sf or tibble, mirroring current behavior
}
\keyword{internal}
