% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_aggregate.R
\name{ducksf_compile_aggs}
\alias{ducksf_compile_aggs}
\title{Compile aggregation view for AW/IW (area/length/count kernels)}
\usage{
ducksf_compile_aggs(
  con,
  ex_vars = character(0),
  in_vars = character(0),
  measure = c("area", "length", "count"),
  overlap_view = "overlap",
  source_proj_view = "source_proj",
  denom_sid_view = "total_by_sid",
  denom_tid_view = "total_by_tid",
  denom_sid_alias = NULL,
  denom_tid_alias = NULL,
  out_view = "interpolated_all",
  tid_col = "tid"
)
}
\description{
Compile aggregation view for AW/IW (area/length/count kernels)
}
\keyword{internal}
