% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_views.R
\name{ducksf_create_norm_view}
\alias{ducksf_create_norm_view}
\title{Normalize a raw table into <id, geom, ...vars> with canonical names}
\usage{
ducksf_create_norm_view(
  con,
  raw_tbl,
  out_view,
  id_col,
  out_id = c("sid", "tid"),
  geom_col_raw = "geometry",
  geom_alias = "geom",
  keep_cols = character(0)
)
}
\description{
Normalize a raw table into <id, geom, ...vars> with canonical names
}
\keyword{internal}
