% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ducksf_args_check.R
\name{ducksf_handle_id_conflict}
\alias{ducksf_handle_id_conflict}
\title{Handle sid/tid same-name conflict on the target}
\usage{
ducksf_handle_id_conflict(target_sf, tidQN, sidQN)
}
\description{
Handle sid/tid same-name conflict on the target
}
\keyword{internal}
