% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ducksf_args_check.R
\name{ducksf_interpolate_aw_plan_vars}
\alias{ducksf_interpolate_aw_plan_vars}
\title{Validate/select variables, enforce numeric, apply na.rm, infer type}
\usage{
ducksf_interpolate_aw_plan_vars(source_sf, extensive, intensive, na.rm = FALSE)
}
\description{
Validate/select variables, enforce numeric, apply na.rm, infer type
}
\keyword{internal}
