% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_views.R
\name{ducksf_project_view}
\alias{ducksf_project_view}
\title{Project a view's geometry into a new view (or pass-through if CRS matches)}
\usage{
ducksf_project_view(
  con,
  in_view,
  out_view,
  from_crs_lit,
  to_crs_lit,
  same_crs = FALSE,
  geom_col = "geom"
)
}
\description{
Project a view's geometry into a new view (or pass-through if CRS matches)
}
\keyword{internal}
