% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-sf.R
\name{st_read.duckdb_connection}
\alias{st_read.duckdb_connection}
\alias{st_read}
\title{Read simple features or layers from a duckdb connection}
\usage{
st_read(dsn, layer, ...)
}
\description{
Read simple features or layers from a duckdb connection
}
